requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.27
asyncpg==0.29.0
python-dotenv==1.0.1
openai==1.12.0
apscheduler==3.10.4
requests==2.31.0
httpx==0.27.0